+++ R/idotplot.R
@@ -5,9 +5,10 @@
 #'
 #' Creates an interactive graph of phenotypes vs genotypes at a marker.
 #'
-#' @param group Vector of groups of individuals (e.g., a genotype)
+#' @param x Vector of groups of individuals (e.g., a genotype)
 #' @param y Numeric vector (e.g., a phenotype)
 #' @param indID Optional vector of character strings, shown with tool tips
+#' @param group Optional vector of categories for coloring points
 #' @param chartOpts A list of options for configuring the chart.  Each
 #'   element must be named using the corresponding option.
 #' @param digits Round data to this number of significant digits
@@ -30,34 +31,37 @@
 #'
 #' @export
 idotplot <-
-function(group, y, indID=NULL, chartOpts=NULL, digits=5)
+function(x, y, indID=NULL, group=NULL, chartOpts=NULL, digits=5)
 {
-    stopifnot(length(group) == length(y))
-    if(is.null(indID))
-        indID <- get_indID(length(group), names(group), names(y))
-    stopifnot(length(indID) == length(group))
-    indID <- as.character(indID)
-    group_levels <- sort(unique(group))
+    stopifnot(length(x) == length(y))
+    if(is.null(group)) group <- rep(1, length(x))
+    stopifnot(length(group) == length(x))
     group <- group2numeric(group)
-    names(y) <- NULL # strip off the names
+    if(is.null(indID))
+        indID <- get_indID(length(x), names(x), names(y), names(group))
+    stopifnot(length(indID) == length(x))
+    indID <- as.character(indID)
+    if(is.factor(x)) x_levels <- levels(x)
+    else x_levels <- sort(unique(x))
+    x <- group2numeric(x)
 
-    chartOpts <- add2chartOpts(chartOpts, ylab="y", title="", xlab="group")
+    # strip off the names
+    names(x) <- NULL
+    names(y) <- NULL
+    names(indID) <- NULL
+    names(group) <- NULL
 
-    # a bit of contortion with the data, to reuse iplotPXG
-    x <- list(data=list(geno=matrix(group, nrow=1),
-                        pheno=y,
-                        chrByMarkers=list(group="un"),
-                        indID=indID,
-                        chrtype=list(un="A"),
-                        genonames=list(A=group_levels)),
-              chartOpts=chartOpts)
+    chartOpts <- add2chartOpts(chartOpts, ylab="y", title="", xlab="group",
+                               xcategories=seq(along=x_levels), xcatlabels=x_levels)
+
+    x <- list(data=list(x=x, y=y, indID=indID, group=group), chartOpts=chartOpts)
     if(!is.null(digits))
         attr(x, "TOJSON_ARGS") <- list(digits=digits)
 
     defaultAspect <- 1 # width/height
     browsersize <- getPlotSize(defaultAspect)
 
-    htmlwidgets::createWidget("iplotPXG", x,
+    htmlwidgets::createWidget("idotplot", x,
                               width=chartOpts$width,
                               height=chartOpts$height,
                               sizingPolicy=htmlwidgets::sizingPolicy(
@@ -66,4 +70,17 @@ function(group, y, indID=NULL, chartOpts=NULL, digits=5)
                                   knitr.defaultWidth=1000,
                                   knitr.defaultHeight=1000/defaultAspect),
                               package="qtlcharts")
+}
+
+#' @rdname qtlcharts-shiny
+#' @export
+idotplot_output <- function(outputId, width="100%", height="530") {
+    htmlwidgets::shinyWidgetOutput(outputId, "idotplot", width, height, package="qtlcharts")
+}
+
+#' @rdname qtlcharts-shiny
+#' @export
+idotplot_render <- function(expr, env=parent.frame(), quoted=FALSE) {
+    if(!quoted) { expr <- substitute(expr) } # force quoted
+    htmlwidgets::shinyRenderWidget(expr, idotplot_output, env, quoted=TRUE)
 }